another commit affraid to lose all my work, the new problem in the windows os is a nightmare

+++ database/migrations/2024_06_25_190802_create_users_table.php
@@ -31,6 +31,8 @@ return new class extends Migration
             $table->timestamp('created_at')->nullable();
         });
 
+        if (config('session.driver') === "database") return;
+
         Schema::create('sessions', function (Blueprint $table) {
             $table->string('id')->primary();
             $table->foreignId('user_id')->nullable()->index();
@@ -48,6 +50,8 @@ return new class extends Migration
     {
         Schema::dropIfExists('users');
         Schema::dropIfExists('password_reset_tokens');
+        if (config('session.driver') === "database") return;
+
         Schema::dropIfExists('sessions');
     }
 };
